# NorthWindApp.WebApiTest/swagger.json
{
  "x-generator": "NSwag v13.7.0.0 (NJsonSchema v10.1.24.0 (Newtonsoft.Json v11.0.0.0))",
  "swagger": "2.0",
  "info": {
    "title": "My Title",
    "version": "1.0.0"
  },
  "host": "localhost:44384",
  "schemes": [
    "https"
  ],
  "paths": {
    "/api/category": {
      "get": {
        "tags": [
          "CategoryApi"
        ],
        "operationId": "CategoryApi_Get",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "404": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      }
    },
    "/api/category/{id}": {
      "get": {
        "tags": [
          "CategoryApi"
        ],
        "operationId": "CategoryApi_Get2",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "404": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      }
    },
    "/api/category/image/{id}": {
      "get": {
        "tags": [
          "CategoryApi"
        ],
        "operationId": "CategoryApi_GetImage",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "404": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      },
      "put": {
        "tags": [
          "CategoryApi"
        ],
        "operationId": "CategoryApi_UpdateImage",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "type": "file",
            "name": "image",
            "in": "formData",
            "x-nullable": true
          }
        ],
        "responses": {
          "404": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      }
    },
    "/api/product": {
      "get": {
        "tags": [
          "ProductApi"
        ],
        "operationId": "ProductApi_Get",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "404": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      },
      "post": {
        "tags": [
          "ProductApi"
        ],
        "operationId": "ProductApi_Create",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "name": "productApi",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ProductApi"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      }
    },
    "/api/product/{id}": {
      "get": {
        "tags": [
          "ProductApi"
        ],
        "operationId": "ProductApi_Get2",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "404": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      },
      "put": {
        "tags": [
          "ProductApi"
        ],
        "operationId": "ProductApi_Update",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "name": "productApi",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ProductApi"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "ProductApi"
        ],
        "operationId": "ProductApi_Delete",
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "ProblemDetails": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "status": {
          "type": "integer",
          "format": "int32"
        },
        "detail": {
          "type": "string"
        },
        "instance": {
          "type": "string"
        },
        "extensions": {
          "type": "object",
          "additionalProperties": {}
        }
      }
    },
    "ProductApi": {
      "allOf": [
        {
          "$ref": "#/definitions/Resource"
        },
        {
          "type": "object",
          "required": [
            "id",
            "name",
            "discontinued"
          ],
          "properties": {
            "id": {
              "type": "integer",
              "format": "int32"
            },
            "name": {
              "type": "string",
              "minLength": 1
            },
            "supplierId": {
              "type": "integer",
              "format": "int32"
            },
            "categoryId": {
              "type": "integer",
              "format": "int32"
            },
            "quantityPerUnit": {
              "type": "string"
            },
            "unitPrice": {
              "type": "number",
              "format": "decimal",
              "maximum": 999.99,
              "minimum": 0.0
            },
            "unitsInStock": {
              "type": "integer"
            },
            "unitsOnOrder": {
              "type": "integer"
            },
            "reorderLevel": {
              "type": "integer"
            },
            "discontinued": {
              "type": "boolean"
            }
          }
        }
      ]
    },
    "Resource": {
      "type": "object",
      "x-abstract": true,
      "properties": {
        "href": {
          "type": "string"
        }
      }
    }
  }
}